@@ -15,17 +15,16 @@ apply plugin: 'net.minecraftforge.gradle'
 apply plugin: 'org.spongepowered.mixin'
 
 group = 'com.voila'
-version = '1.0'
+version = '1.1'
 
 java {
     archivesBaseName = 'forgetest'
-    toolchain.languageVersion = JavaLanguageVersion.of(8)
+    toolchain.languageVersion = JavaLanguageVersion.of(17)
 }
 
 compileJava {
     compileJava.options.encoding = 'UTF-8'
 }
-
 
 minecraft {
     // The mappings can be changed at any time and must be in the following format.
