@@ -3,62 +3,62 @@ package com.voila.forge;
 import com.mojang.brigadier.*;
 import com.mojang.brigadier.exceptions.*;
 import com.mojang.serialization.*;
+import net.minecraft.client.multiplayer.*;
 import net.minecraft.client.particle.*;
-import net.minecraft.client.world.*;
+import net.minecraft.core.particles.*;
 import net.minecraft.network.*;
-import net.minecraft.particles.*;
 import net.minecraftforge.api.distmarker.*;
 import org.apache.logging.log4j.*;
 
 import javax.annotation.*;
 
-public class DamageParticle extends SpriteTexturedParticle {
+public class DamageParticle extends TextureSheetParticle {
 	int amount;
 	int ma = 15;
 	static Logger Log = LogManager.getLogger();
-	IAnimatedSprite sprite;
+	SpriteSet sprite;
 
-	protected DamageParticle(ClientWorld world, double x, double y, double z, double motionX, double motionY, double motionZ, int amount, IAnimatedSprite as){
-		super(world, x, y, z, motionX, motionY, motionZ);
-		maxAge = 2048;
-		canCollide = false;
+	protected DamageParticle(ClientLevel world, double x, double y, double z, double xd, double yd, double zd, int amount, SpriteSet as){
+		super(world, x, y, z, xd, yd, zd);
+		lifetime = 2048;
+		hasPhysics = false;
 		this.amount = Math.max(amount, 0);
 		sprite = as;
 		setSprite(sprite.get(amount, 2048));
-		if(this.motionY < 0)
-			this.motionY = -this.motionY;
+		if(this.yd < 0)
+			this.yd = -this.yd;
 
 	}
 
 	@Override
 	public void tick(){
 		super.tick();
-		this.motionX *= 0.9;
-		this.motionY *= 0.7;
-		this.motionZ *= 0.9;
-		this.motionY -= 0.02;
+		this.xd *= 0.9;
+		this.yd *= 0.7;
+		this.zd *= 0.9;
+		this.yd -= 0.02;
 		setSprite(sprite.get(amount, 2048));
 		if(age > ma)
-			setExpired();
+			remove();
 	}
 
 	@Override
-	public IParticleRenderType getRenderType(){
-		return IParticleRenderType.PARTICLE_SHEET_LIT;
+	public ParticleRenderType getRenderType(){
+		return ParticleRenderType.PARTICLE_SHEET_LIT;
 	}
 
-	public static class DamageParticleData extends ParticleType<DamageParticleData> implements IParticleData {
+	public static class DamageParticleData extends ParticleType<DamageParticleData> implements ParticleOptions {
 		int amount;
-		public static final IDeserializer<DamageParticleData> deserializer = new IDeserializer<DamageParticleData>() {
+		public static final Deserializer<DamageParticleData> deserializer = new Deserializer<DamageParticleData>() {
 			@Override
-			public DamageParticleData deserialize(ParticleType<DamageParticleData> particleTypeIn, StringReader reader) throws CommandSyntaxException{
+			public DamageParticleData fromCommand(ParticleType<DamageParticleData> particleTypeIn, StringReader reader) throws CommandSyntaxException{
 				reader.expect(' ');
 				int amount = reader.readInt();
 				return new DamageParticleData(amount);
 			}
 
 			@Override
-			public DamageParticleData read(ParticleType<DamageParticleData> particleTypeIn, PacketBuffer buffer){
+			public DamageParticleData fromNetwork(ParticleType<DamageParticleData> particleTypeIn, FriendlyByteBuf buffer){
 				return new DamageParticleData(buffer.readInt());
 			}
 		};
@@ -78,17 +78,17 @@ public class DamageParticle extends SpriteTexturedParticle {
 		}
 
 		@Override
-		public void write(PacketBuffer buffer){
+		public void writeToNetwork(FriendlyByteBuf buffer){
 			buffer.writeInt(this.amount);
 		}
 
 		@Override
-		public String getParameters(){
+		public String writeToString(){
 			return String.format("%s %d", this.getType().getRegistryName(), this.amount);
 		}
 
 		@Override
-		public Codec<DamageParticleData> func_230522_e_(){
+		public Codec<DamageParticleData> codec(){
 			return Codec.unit(() -> DamageParticleData.this);
 		}
 
@@ -98,16 +98,16 @@ public class DamageParticle extends SpriteTexturedParticle {
 	}
 
 	@OnlyIn(Dist.CLIENT)
-	public static class Factory implements IParticleFactory<DamageParticleData> {
-		private final IAnimatedSprite sprites;
+	public static class Factory implements ParticleProvider<DamageParticleData> {
+		private final SpriteSet sprites;
 
-		public Factory(IAnimatedSprite sprite){
+		public Factory(SpriteSet sprite){
 			this.sprites = sprite;
 		}
 
 		@Nullable
 		@Override
-		public Particle makeParticle(DamageParticleData typeIn, ClientWorld worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed){
+		public Particle createParticle(DamageParticleData typeIn, ClientLevel worldIn, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed){
 			DamageParticle particle = new DamageParticle(worldIn, x, y, z, xSpeed, ySpeed, zSpeed, typeIn.getAmount(), sprites);
 			return particle;
 		}
