@@ -1,9 +1,15 @@
 package com.voila.forge;
 
+import net.minecraft.client.*;
+
 public interface IMinecraft {
-	void pickBlock();
+	void pick();
 
 	void use();
 
 	void attack();
+
+	User getSession();
+
+	void setSession(User se);
 }
