@@ -1,11 +1,10 @@
 package com.voila.forge;
 
-import com.mojang.blaze3d.matrix.*;
+import com.mojang.blaze3d.vertex.*;
 import net.minecraft.client.*;
-import net.minecraft.client.gui.screen.*;
-import net.minecraft.client.gui.widget.button.*;
-import net.minecraft.client.gui.widget.list.*;
-import net.minecraft.util.text.*;
+import net.minecraft.client.gui.components.*;
+import net.minecraft.client.gui.screens.*;
+import net.minecraft.network.chat.*;
 
 import java.io.*;
 
@@ -14,28 +13,28 @@ public class SelectScriptScreen extends Screen {
 	Button doneButton;
 
 	public SelectScriptScreen(){
-		super(new TranslationTextComponent("title." + Forgetest.ID + ".selectScript"));
+		super(new TranslatableComponent("title." + Forgetest.ID + ".selectScript"));
 	}
 
 	@Override
 	protected void init(){
 		super.init();
 		doneButton = new Button((width - 150) / 2, height - 32, 150, 20,
-			new TranslationTextComponent("title." + Forgetest.ID + ".save"), (button) ->
-			Minecraft.getInstance().displayGuiScreen(null));
+			new TranslatableComponent("title." + Forgetest.ID + ".save"), (button) ->
+			Minecraft.getInstance().setScreen(null));
 		list = new ScriptList();
 
-		children.add(list);
-		addButton(doneButton);
+		addWidget(list);
+		addRenderableWidget(doneButton);
 
 
 	}
 
 	@Override
-	public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks){
+	public void render(PoseStack matrixStack, int mouseX, int mouseY, float partialTicks){
 		super.render(matrixStack, mouseX, mouseY, partialTicks);
 		renderBackground(matrixStack);
-		font.drawText(matrixStack, title, (width - font.getStringPropertyWidth(title)) / 2.0f, 20, 0xffffff);
+		font.draw(matrixStack, title, (width - font.width(title)) / 2.0f, 20, 0xffffff);
 		list.render(matrixStack, mouseX, mouseY, partialTicks);
 		doneButton.render(matrixStack, mouseX, mouseY, partialTicks);
 	}
@@ -45,13 +44,13 @@ public class SelectScriptScreen extends Screen {
 		return false;
 	}
 
-	class ScriptList extends ExtendedList<ScriptList.ScriptEntry> {
+	class ScriptList extends ObjectSelectionList<ScriptList.ScriptEntry> {
 
 		public ScriptList(){
 			super(Minecraft.getInstance(), SelectScriptScreen.this.width, SelectScriptScreen.this.height,
 				48, SelectScriptScreen.this.height - 20 - 32 - 30, 20);
-			this.func_244605_b(false);
-			this.func_244606_c(false);
+			setRenderBackground(false);
+			setRenderTopAndBottom(false);
 
 			File scriptDir = new File("scripts");
 			scriptDir.mkdirs();
@@ -64,13 +63,13 @@ public class SelectScriptScreen extends Screen {
 
 		}
 
-		class ScriptEntry extends ExtendedList.AbstractListEntry<ScriptEntry> {
+		class ScriptEntry extends ObjectSelectionList.Entry<ScriptEntry> {
 
 			String name;
 
 			@Override
-			public void render(MatrixStack matrixStack, int index, int top, int left, int width, int height, int mouseX, int mouseY, boolean isMouseOver, float partialTicks){
-				font.drawString(matrixStack, name, left + 2, top + 3, 0xffffff);
+			public void render(PoseStack matrixStack, int index, int top, int left, int width, int height, int mouseX, int mouseY, boolean isMouseOver, float partialTicks){
+				font.draw(matrixStack, name, left + 2, top + 3, 0xffffff);
 			}
 
 			public ScriptEntry(String filename){
@@ -85,6 +84,10 @@ public class SelectScriptScreen extends Screen {
 			}
 
 
+			@Override
+			public Component getNarration(){
+				return TextComponent.EMPTY;
+			}
 		}
 
 	}
